+++ src/components/victory-area/victory-area.js
@@ -14,7 +14,7 @@ const defaultStyles = {
     fill: "#756f6a"
   },
   labels: {
-    fontSize: 12,
+    fontSize: 13,
     padding: 4,
     fill: "black"
   }
@@ -347,7 +347,7 @@ export default class VictoryArea extends React.Component {
     this.baseProps = AreaHelpers.getBaseProps(newProps, defaultStyles);
   }
 
-  renderArea(props) {
+  renderData(props) {
     const { dataComponent, labelComponent, sharedEvents } = props;
     const getSharedEventState = sharedEvents && isFunction(sharedEvents.getEventState) ?
       sharedEvents.getEventState : () => undefined;
@@ -356,8 +356,8 @@ export default class VictoryArea extends React.Component {
       {},
       this.getEventState("all", "data"),
       getSharedEventState("all", "data"),
-      this.baseProps.all.data,
-      dataComponent.props
+      dataComponent.props,
+      this.baseProps.all.data
     );
     const areaComponent = React.cloneElement(dataComponent, Object.assign(
       {}, dataProps, {events: Events.getPartialEvents(dataEvents, "all", dataProps)}
@@ -367,8 +367,8 @@ export default class VictoryArea extends React.Component {
         {},
         this.getEventState("all", "labels"),
         getSharedEventState("all", "labels"),
-        this.baseProps.all.labels,
-        labelComponent.props
+        labelComponent.props,
+        this.baseProps.all.labels
       );
     if (labelProps && labelProps.text) {
       const labelEvents = this.getEvents(props, "labels", "all");
@@ -386,12 +386,14 @@ export default class VictoryArea extends React.Component {
   }
 
   render() {
-    if (this.props.animate) {
+    const { animate, style, width, height, standalone, containerComponent } = this.props;
+
+    if (animate) {
       const whitelist = [
         "data", "domain", "height", "padding", "style", "width"
       ];
       return (
-        <VictoryTransition animate={this.props.animate} animationWhitelist={whitelist}>
+        <VictoryTransition animate={animate} animationWhitelist={whitelist}>
           <VictoryArea {...this.props}/>
         </VictoryTransition>
       );
@@ -399,20 +401,19 @@ export default class VictoryArea extends React.Component {
 
     const styleObject = this.props.theme && this.props.theme.area ? this.props.theme.area
     : defaultStyles;
-    const style = Helpers.getStyles(
-      this.props.style,
-      styleObject,
-      "auto",
-      "100%"
+
+    const baseStyles = Helpers.getStyles(style, styleObject, "auto", "100%");
+
+    const group = (
+      <g role="presentation" style={baseStyles.parent}>
+        {this.renderData(this.props)}
+      </g>
     );
-    const group = <g role="presentation" style={style.parent}>{this.renderArea(this.props)}</g>;
-    return this.props.standalone ?
+
+    return standalone ?
       React.cloneElement(
-        this.props.containerComponent,
-        Object.assign({
-          height: this.props.height,
-          width: this.props.width,
-          style: style.parent}, this.props.containerComponent.props),
+        containerComponent,
+        Object.assign({ height, width, style: baseStyles.parent}, containerComponent.props),
         group) :
       group;
   }