+++ src/components/victory-bar/victory-bar.js
@@ -19,7 +19,7 @@ const defaultStyles = {
     opacity: 1
   },
   labels: {
-    fontSize: 12,
+    fontSize: 13,
     padding: 4,
     fill: "black"
   }
@@ -362,8 +362,8 @@ export default class VictoryBar extends React.Component {
         {key: `bar-${key}`},
         this.getEventState(key, "data"),
         getSharedEventState(key, "data"),
-        this.baseProps[key].data,
-        dataComponent.props
+        dataComponent.props,
+        this.baseProps[key].data
       );
       const barComponent = React.cloneElement(dataComponent, Object.assign(
         {}, dataProps, {events: Events.getPartialEvents(dataEvents, key, dataProps)}
@@ -372,8 +372,8 @@ export default class VictoryBar extends React.Component {
         {key: `bar-label-${key}`},
         this.getEventState(key, "labels"),
         getSharedEventState(key, "labels"),
-        this.baseProps[key].labels,
-        labelComponent.props
+        labelComponent.props,
+        this.baseProps[key].labels
       );
       if (labelProps && labelProps.text) {
         const labelEvents = this.getEvents(props, "labels", key);
@@ -392,30 +392,34 @@ export default class VictoryBar extends React.Component {
   }
 
   render() {
-    // If animating, return a `VictoryAnimation` element that will create
-    // a new `VictoryBar` with nearly identical props, except (1) tweened
-    // and (2) `animate` set to null so we don't recurse forever.
-    if (this.props.animate) {
+    const {animate, style, standalone, containerComponent, height, width} = this.props;
+
+    if (animate) {
       const whitelist = [
         "data", "domain", "height", "padding", "style", "width"
       ];
       return (
-        <VictoryTransition animate={this.props.animate} animationWhitelist={whitelist}>
+        <VictoryTransition animate={animate} animationWhitelist={whitelist}>
           <VictoryBar {...this.props}/>
         </VictoryTransition>
       );
     }
+
     const styleObject = this.props.theme && this.props.theme.bar ? this.props.theme.bar
     : defaultStyles;
-    const style = Helpers.getStyles(this.props.style, styleObject, "auto", "100%");
-    const group = <g style={style.parent} role="presentation">{this.renderData(this.props)}</g>;
-    return this.props.standalone ?
+
+    const baseStyles = Helpers.getStyles(style, styleObject, "auto", "100%");
+
+    const group = (
+      <g role="presentation" style={baseStyles.parent}>
+        {this.renderData(this.props)}
+      </g>
+    );
+
+    return standalone ?
       React.cloneElement(
-        this.props.containerComponent,
-        Object.assign({
-          height: this.props.height,
-          width: this.props.width,
-          style: style.parent}, this.props.containerComponent.props),
+        containerComponent,
+        Object.assign({ height, width, style: baseStyles.parent}, containerComponent.props),
         group) :
       group;
   }