+++ src/components/victory-group/victory-group.js
@@ -192,8 +192,9 @@ export default class VictoryGroup extends React.Component {
     labelComponent: PropTypes.element,
     /**
      * The offset prop derermines the number of pixels each element in a group should
-     * be offset from the others. In the case of groups of bars, this number should
-     * be equal to the width of the bar plus the desired spacing between bars.
+     * be offset from its original position on the independent axis. In the case of
+     * groups of bars, this number should be equal to the width of the bar plus
+     * the desired spacing between bars.
      */
     offset: PropTypes.number,
     /**
@@ -305,9 +306,11 @@ export default class VictoryGroup extends React.Component {
       x: Scale.getScaleFromProps(props, "x") || Scale.getDefaultScale(),
       y: Scale.getScaleFromProps(props, "y") || Scale.getDefaultScale()
     };
+    const xScale = baseScale.x.domain(domain.x).range(range.x);
+    const yScale = baseScale.y.domain(domain.y).range(range.y);
     const scale = {
-      x: baseScale.x.domain(domain.x).range(range.x),
-      y: baseScale.y.domain(domain.y).range(range.y)
+      x: horizontal ? yScale : xScale,
+      y: horizontal ? xScale : yScale
     };
     const categories = {
       x: Wrapper.getCategories(props, "x"),
@@ -317,20 +320,25 @@ export default class VictoryGroup extends React.Component {
     return {datasets, categories, range, domain, horizontal, scale, style, colorScale};
   }
 
-  pixelsToValue(pixels, axis, calculatedProps) {
-    if (pixels === 0) {
+  pixelsToValue(props, axis, calculatedProps) {
+    if (props.offset === 0) {
       return 0;
     }
-    const domain = calculatedProps.domain[axis];
-    const range = calculatedProps.range[axis];
+    const childComponents = React.Children.toArray(props.children);
+    const horizontalChildren = childComponents.some((child) => child.props.horizontal);
+    const horizontal = props && props.horizontal || horizontalChildren.length > 0;
+    const otherAxis = axis === "x" ? "y" : "x";
+    const currentAxis = horizontal ? otherAxis : axis;
+    const domain = calculatedProps.domain[currentAxis];
+    const range = calculatedProps.range[currentAxis];
     const domainExtent = Math.max(...domain) - Math.min(...domain);
     const rangeExtent = Math.max(...range) - Math.min(...range);
-    return domainExtent / rangeExtent * pixels;
+    return domainExtent / rangeExtent * props.offset;
   }
 
   getXO(props, calculatedProps, datasets, index) { // eslint-disable-line max-params
     const center = (datasets.length - 1) / 2;
-    const totalWidth = this.pixelsToValue(props.offset, "x", calculatedProps);
+    const totalWidth = this.pixelsToValue(props, "x", calculatedProps);
     return (index - center) * totalWidth;
   }
 
@@ -347,7 +355,6 @@ export default class VictoryGroup extends React.Component {
       height: props.height,
       width: props.width,
       padding: Helpers.getPadding(props),
-      labelComponent: props.labelComponent,
       standalone: false,
       theme: props.theme,
       categories,
@@ -374,10 +381,11 @@ export default class VictoryGroup extends React.Component {
       const xOffset = this.getXO(props, calculatedProps, datasets, index);
       const data = datasets[index].map((datum) => Object.assign({}, datum, {xOffset}));
       const style = Wrapper.getChildStyle(child, index, calculatedProps);
+      const labels = props.labels ? this.getLabels(props, datasets, index) : child.props.labels;
       return React.cloneElement(child, Object.assign({
         animate: getAnimationProps(props, child, index),
         key: index,
-        labels: this.getLabels(props, datasets, index) || child.props.labels,
+        labels,
         labelComponent: props.labelComponent || child.props.labelComponent,
         style,
         data,
@@ -411,13 +419,12 @@ export default class VictoryGroup extends React.Component {
         {newChildren}
       </g>
     );
-    return this.props.standalone ?
+
+    const { standalone, width, height, containerComponent } = this.props;
+    return standalone ?
       React.cloneElement(
-        this.props.containerComponent,
-        Object.assign({
-          height: this.props.height,
-          width: this.props.width,
-          style: style.parent}, this.props.containerComponent.props),
+        containerComponent,
+        Object.assign({ height, width, style: style.parent}, containerComponent.props),
         group) :
       group;
   }